Unit Squads, Hero's, Etc.
Added a Menu object for use in the command section.
Removed PlayerCommand.cs as it's functions are depreciated.
Created the start of Generals and Champions via a bool option in the
UnitControl.

TO DO: Add Menu functionality, Create Enum values for unit status'
(Stunned, sleep, etc), Add special attacks for General's and Champions